+++ stdlib/public/Concurrency/CMakeLists.txt
@@ -104,7 +104,6 @@ add_swift_target_library(swift_Concurrency ${SWIFT_STDLIB_LIBRARY_BUILD_TYPES} I
   AsyncThrowingFlatMapSequence.swift
   AsyncThrowingMapSequence.swift
   AsyncThrowingPrefixWhileSequence.swift
-  UncheckedContinuation.swift
   GlobalActor.swift
   MainActor.swift
   PartialAsyncTask.swift
